:recycle: refactor : 게시글 영속화 객체에서 도메인 객체로 변환하는 부분 별도의 매퍼 파일로 이동
